requirements.txt
numpy
scipy
absl-py
matplotlib
scikit-image  #  for implicit function of geometric primitives
